Teams
“Dictator” Superclass ✓
Creating teams in “Teams” class ✓
