+++ Erdogan.java
@@ -1,0 +1,4 @@
+package Dictators;
+
+public class Erdogan extends Dictator {
+}
